added order and selection options
internal get_samples() function added to obtain samples from multiindex-columns keeping their order. This is used in the following functions.

tir.summarize(sort_samples = False) is a new parameter, by the way the 'label' subcolumn is now arranged within the other columns of an sample.
tir.concatenate(sample_order = []) is a new parameter, that allows to specify an output order of the sample columns and its selection.